test: migrate web site controller tests to Testcontainers
Replace the mock-based web site controller unit tests with container-backed integration tests and remove the superseded unit test file.
